% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{amphi_lifehist}
\alias{amphi_lifehist}
\title{Life-history traits of European amphibians}
\format{
A \code{data.frame} with 86 observations and 256 variables.
}
\source{
Cite this dataset as: \itemize{\item Trochet A, Moulherat S, Calvez O, Stevens V, Clobert J, Schmeller D (2014) A database of life-history
traits of European amphibians. Biodiversity Data Journal 2: e4123. \href{http://doi.org/10.3897/BDJ.2.e4123}{10.3897/BDJ.2.e4123}.}
}
\usage{
data(amphi_lifehist)
}
\description{
A database of life-history traits of European amphibians
}
\details{
In the current context of climate change and landscape fragmentation, efficient
conservation strategies require the explicit consideration of life history traits. This is 
particularly true for amphibians, which are highly threatened worldwide, composed by more 
than 7400 species, which is constitute one of the most species-rich vertebrate groups. The
collection of information on life history traits is difficult due to the ecology of species and
remoteness of their habitats. It is therefore not surprising that our knowledge is limited, and
missing information on certain life history traits are common for in this species group. We
compiled data on amphibian life history traits from literature in an extensive database with
morphological and behavioral traits, habitat preferences and movement abilities for 86
European amphibian species (50 Anuran and 36 Urodela species). When it were available,
we reported data for males, females, juveniles and tadpoles. Our database may serve as
an important starting point for further analyses regarding amphibian conservation.
}
