% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{amphibio}
\alias{amphibio}
\title{A global database for amphibian ecological traits}
\format{
A \code{data.frame} with 6776 observations and 39 variables.
}
\source{
Cite this dataset as: \itemize{\item Oliveira, B.F., Sao-Pedro, V.A., Santos-Barrera, G., Penone,
C. & Costa, G.C. (2017). AmphiBIO, a global database for amphibian ecological traits.
Scientific Data. \href{https://www.nature.com/articles/sdata2017123}{10.1038/sdata.2017.123}.}

Please also cite the data repository on figshare: \itemize{\item Oliveira, Brunno Freire; Sao-Pedro, 
Vinicius Avelar; Santos-Barrera, Georgina; Penone, Caterina; C. Costa, Gabriel (2017):
AmphiBIO_v1. \href{https://doi.org/10.6084/m9.figshare.4644424.v5}{10.6084/m9.figshare.4644424.v5}.}
}
\usage{
data(amphibio)
}
\description{
A comprehensive database of natural history traits for amphibians worldwide.
}
\details{
Current ecological and evolutionary research are increasingly moving from species- to trait-based approaches
because traits provide a stronger link to organism's function and fitness. Trait databases covering a large number of species
are becoming available, but such data remains scarce for certain groups. Amphibians are among the most diverse vertebrate groups
on Earth, and constitute an abundant component of major terrestrial and freshwater ecosystems. They are also facing rapid population
declines worldwide, which is likely to affect trait composition in local communities, thereby impacting ecosystem processes
and services. In this context, we introduce AmphiBIO, a comprehensive database of natural history traits for amphibians worldwide.
The database releases information on 17 traits related to ecology, morphology and reproduction features of amphibians. We compiled data
from more than 1500 literature sources, and for more than 6500 species of all orders, 61 families and 531 genera. 
This database has the potential to allow unprecedented large-scale analyses in ecology, evolution and conservation of amphibians.
}
\section{Measures}{

\itemize{
    \item id, Order, Family, Genus, Species,
    \item Fos, Ter, Aqu, Arb, Leaves, Flowers, Seeds, Fruits,
    \item Arthro, Vert, Diu, Noc, Crepu, Wet_warm, Wet_cold, Dry_warm, Dry_cold,
    \item Body_mass_g, Age_at_maturity_min_y, Age_at_maturity_max_y, Body_size_mm,
    \item Size_at_maturity_min_mm, Size_at_maturity_max_mm, Longevity_max_y, Litter_size_min_n, Litter_size_max_n,
    \item Reproductive_output_y, Offspring_size_min_mm, Offspring_size_max_mm, Dir, Lar, Viv, OBS,
    \item scientificNameStd
  }
}

