% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{an_age}
\alias{an_age}
\title{Database of Animal Ageing and Longevity}
\format{
A \code{data.frame} with 4219 observations and 32 variables.
}
\source{
When using this data, please cite the original publication:
  \itemize{\item  Tacutu, R., Craig, T., Budovsky, A., Wuttke, D., Lehmann, G., Taranukha, D., Costa, J., Fraifeld, V. E., de Magalhaes, J. P. (2013)
  "Human Ageing Genomic Resources: Integrated databases and tools for the biology and genetics of ageing." 
  Nucleic Acids Research 41(D1):D1027-D1033.}
}
\usage{
data(an_age)
}
\description{
Data on animal ageing and longevity
}
\details{
The data set comprises data on the age and longevity of multiple animal species
}
