% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{anuran_morpho}
\alias{anuran_morpho}
\title{Database of Animal Ageing and Longevity}
\format{
A \code{data.frame} with 4623 observations and 29 variables.
}
\source{
When using this data, please cite the original publication:
   \itemize{\item  Mendoza-Henao et al. (2019) 
   "A morphological database for Colombian anuran species from conservation‐priority ecosystems." 
   Ecology 100.}
}
\usage{
data(anuran_morpho)
}
\description{
Data on animal ageing and longevity
}
\details{
The data set comprises morphological data for Colombian anuran species.
}
