% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{arthropods}
\alias{arthropods}
\title{Functional Arthropod Traits}
\format{
A \code{data.frame} with 1230 observations and 19 variables.
}
\source{
When using this data, please cite the original publication:

  \itemize{\item  Gossner MM, Simons NK, Achtziger R, Blick T, Dorow WHO,
  Dziock F, Koehler F, Rabitsch W, Weisser WW (2015) A summary of eight traits
  of Coleoptera, Hemiptera, Orthoptera and Araneae, occurring in grasslands
  in Germany. Scientific Data. \url{http://dx.doi.org/10.1038/sdata.2015.13}.}
  
  Additionally, please cite the Dryad data package:
  
  \itemize{\item  Gossner MM, Simons NK, Achtziger R, Blick T, Dorow WHO, Dziock F, Koehler F, 
  Rabitsch W, Weisser WW (2015) Data from: A summary of eight traits of Coleoptera, 
  Hemiptera, Orthoptera and Araneae, occurring in grasslands in Germany. 
  Dryad Digital Repository. \url{http://dx.doi.org/10.5061/dryad.53ds2}.}
}
\usage{
data(arthropods)
}
\description{
A summary of eight traits of Coleoptera, Hemiptera, Orthoptera and Araneae, occurring in grasslands in Germany.
}
\details{
The data set comprises literature trait data of species that were sampled and 
measured in a project within the Biodiversity Exploratories which focuses on the effect of land use 
on arthropod community composition and related processes in three regions of Germany
}
\section{Measures}{


\itemize{ 
    \item Order, Suborder, Family, SpeciesID, Genus, Species, Author
    \item Body_Size, Dispersal_ability,
    \item Feeding_guild, Feeding_guild_short, Feeding_mode, Feeding_specialization, Feeding_tissue, Feeding_plant_part,
    \item Endophagous_lifestyle, Stratum_use, Stratum_use_short, Remark,
    \item scientificNameStd
    }
}

