% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{atlantic_birds}
\alias{atlantic_birds}
\title{Database of Atlantic bird morphological traits}
\format{
A \code{data.frame} with 72483 observations and 83 variables.
}
\source{
When using this data, please cite the original publication:
  \itemize{\item  Rodrigues et al. (2019) 
  "ATLANTIC BIRD TRAITS: a data set of bird morphological traits from the Atlantic forests of South America." 
  Ecology 100.}
}
\usage{
data(atlantic_birds)
}
\description{
Data on morphological traits of Atlantic forest birds
}
\details{
The data set comprises morphological traits of birds from the Atlantic forests of South America
}
