% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{australian_birds}
\alias{australian_birds}
\title{Database of Australian birds}
\format{
A \code{data.frame} with 2056 observations and 230 variables.
}
\source{
When using this data, please cite the original publication:
  \itemize{\item  Garnett et al. (2015)
  Biological, ecological, conservation and legal information for all species and subspecies of Australian bird. 
  Scientific Data.}
}
\usage{
data(australian_birds)
}
\description{
Data on Australian bird traits
}
\details{
The data set comprises biological, ecological, conservation and legal information 
for every species and subspecies of Australian birds
}
