% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{AvianBodySize}
\alias{AvianBodySize}
\title{Database on Avian body sizes}
\format{
A \code{data.frame} with 3802 observations and 46 variables.
}
\source{
When using this data, please cite the original publication:
\itemize{\item  Lislevand et al. 2007 (Ecology) AVIAN BODY SIZES IN RELATION TO FECUNDITY, 
MATING SYSTEM, DISPLAY BEHAVIOR, AND RESOURCE SHARINGe.}.
}
\usage{
data(AvianBodySize)
}
\description{
Data on Avian bod sizes
}
\details{
Species‐specific measurements on male and female body mass, wing length, 
tarsus length, bill length, and tail length from major ornithological text books and 
some other sources covering bird species of Africa, Australia, New Zealand, 
Antarctica, North America, and the western Palearctic. These measurements were 
matched with measures of egg and clutch sizes, and scores of mating system, 
sexual display agility, and the degree of intersexual resource division. We present 
morphometric data ranging from 2350 species (minimum, tail length) to 2979 species (maximum, wing length) 
where measurements for both sexes are known, some additional data where only one sex or unsexed birds 
have been measured, and explanatory data ranging from 1218 species (minimum, display agility) 
to 2603 species (maximum, egg mass). In total, 3769 species from 125 of 146 different bird families are included.
}
