% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{avonet}
\alias{avonet}
\title{Morphological, ecological and geographical data for all birds}
\format{
A \code{data.frame} with 90371 observations and 24 variables.
}
\source{
When using this data, please cite the original publication:
\itemize{\item Tobias et al. 2021 (Ecology Letters) 
AVONET: morphological, ecological and geographical data for all birds.}.
}
\usage{
data(avonet)
}
\description{
Morphological, ecological and geographical data for all birds
}
\details{
The AVONET database contains comprehensive functional trait data for all birds, 
including six ecological variables, eleven continuous morphological traits, 
and information on range size and location. Raw morphological measurements 
are available from 90020 individuals of 11009 extant bird species sampled 
from 181 countries. These data are also summarised as species averages in 
three taxonomic formats, allowing integration with a global phylogeny, 
geographical range maps, IUCN Red List data, and the eBird citizen science 
database. Code to reproduce the analyses and figures presented in 
Tobias et al 2021 "AVONET: morphological, ecological and geographical 
data for all birds" Ecology Letters, is also included.
}
