% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{bird_behav}
\alias{bird_behav}
\title{Foraging behaviour and dietary niche of birds}
\format{
A \code{data.frame} with 9658 observations and 19 variables.
}
\source{
Cite this dataset as: \itemize{\item Tobias J. A., Pigot, A. L. (2019)
Integrating behaviour and ecology into global biodiversity conservation strategies. Phil. Trans. R. Soc. B, 374, 1781.
\url{http://doi.org/10.1098/rstb.2019.0012}.}
}
\usage{
data(bird_behav)
}
\description{
Foraging behaviour and dietary niche of birds
}
\details{
A dataset for 9658 bird species describing IUCN Red List and Threat category, range size and body mass, 
foraging behaviour and diet, ...
}
\section{Measures}{


\itemize{
  \item Genus, Species, Family
  \item RedlistCategory, Threat, LogRangeSize, LogBodyMass, Diet, Foraging, Migration, MatingSystem, NestPlacement, Territoriality, Habitat, 
  IslandDwelling, LogClutchSize, LogNightLights, LogHumanPopulationDensity
  \item scientificNameStd
}
}

