% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{carabids}
\alias{carabids}
\title{Carabid morphological traits}
\format{
A \code{data.frame} with 121 observations and 12 variables.
}
\source{
When using this data, please cite the original publication:

  \itemize{ \item van der Plas F, van Klink R, Manning P, Olff H, Fischer M
  (2017) Sensitivity of functional diversity metrics to sampling intensity.
  Methods in Ecology and Evolution 8(9): 1072-1080.
  \url{https://doi.org/10.1111/2041-210x.12728}.}

  Additionally, please cite the Dryad data package:

  \itemize{ \item van der Plas F, van Klink R, Manning P, Olff H, Fischer M
  (2017) Data from: Sensitivity of functional diversity metrics to sampling
  intensity. Dryad Digital Repository. \url{https://doi.org/10.5061/dryad.1fn46}.}
}
\usage{
data(carabids)
}
\description{
Average body measures of 120 Carabid species occuring in the Netherlands.
}
\details{
Carabid morphological traits. Column names are described below.
}
\section{Measures}{

\itemize{
    \item name_correct = species name,
    \item source_measurement = researcher who performed measurement,
    \item body_length = body length in mm,
    \item antenna_length = antenna length in mm,
    \item metafemur_length = length metafemur in mm,
    \item eyewith_corr = eye width in mm,
    \item note = note,
    \item resid_femur = residual femur length in mm (i.e. residual from 
       linear model in which femur length is explained by body length),
    \item resid_eye = residual eye length in mm (i.e. residual from linear 
       model in which eye length is explained by body length),
    \item resid_antenna = residual antenna length in mm (i.e. residual 
       from linear model in which antenna length is explained by body length),
    \item scientificNameStd
  }
}

