% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{climber}
\alias{climber}
\title{Climatic niche characteristics of European butterflies}
\format{
A \code{data.frame} with 397 observations and 69 variables.
}
\source{
Cite this dataset as \itemize{ \item Schweiger O, Harpke A, Wiemers M, Settele J (2014) 
CLIMBER: Climatic niche characteristics of the butterflies in Europe. ZooKeys 367:65-84. 
\url{http://dx.doi.org/10.3897/zookeys.367.6185}.}
}
\usage{
data(climber)
}
\description{
Climatic niche characteristics of the butterflies in Europe
}
\details{
Detailed information on species' ecological niche characteristics that can be related to declines and 
extinctions is indispensable for a better understanding of the relationship between the occurrence and 
performance of wild species and their environment and, moreover, for an improved assessment of the impacts of
global change. Knowledge on species characteristics such as habitat requirements is already available in the
ecological literature for butterflies, but information about their climatic requirements is still lacking. Here
we present a unique dataset on the climatic niche characteristics of 397 European butterflies representing
91% of the European species. These characteristics were obtained by combining detailed information on 
butterfly distributions in Europe and the corresponding climatic conditions. The presented dataset comprises information for
the position and breadth of the following climatic niche characteristics: mean annual temperature, range
in annual temperature, growing degree days, annual precipitation sum, range in annual precipitation and
soil water content. The climatic niche position is indicated by the median and mean value for each climate
variable across a species' range, accompanied by the 95% confidence interval for the mean and the number
of grid cells used for calculations. Climatic niche breadth is indicated by the standard deviation and the
minimum and maximum values for each climatic variable across a species' range. Database compilation
was based on high quality standards and the data are ready to use for a broad range of applications.
It is already evident that the information provided in this dataset is of great relevance for basic and applied ecology. 
Based on the species temperature index (STI, i.e. the mean temperature value per species),
the community temperature index (CTI, i.e. the average STI value across the species in a community) was
recently adopted as an indicator of climate change impact on biodiversity by the pan-European framework
supporting the Convention on Biological Diversity (Streamlining European Biodiversity Indicators 2010)
and has already been used in several scientific publications. The application potential of this database
ranges from theoretical aspects such as assessments of past niche evolution or analyses of trait interdependencies 
to the very applied aspects of measuring, monitoring and projecting historical, ongoing and
potential future responses to climate change using butterflies as an indicator.
}
