% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{disperse}
\alias{disperse}
\title{Dispersal potential of European aquatic macroinvertebrates}
\format{
A \code{data.frame} with 482 observations and 46 variables.
}
\source{
Cite this dataset as \itemize{ \item Sarremejane, R., Cid, N., Stubbington, R. et al. 
DISPERSE, a trait database to assess the dispersal potential of European aquatic macroinvertebrates. 
Sci Data 7, 386 (2020). \url{https://doi.org/10.1038/s41597-020-00732-7}.}
}
\usage{
data(disperse)
}
\description{
Dataset on the dispersal potential of European aquatic macroinvertebrates
}
\details{
This dataset contains information about the dispersal potential of 
European aquatic macroinvertebrates
}
\section{Measures}{

 
\itemize{
  \item Group, Family, Genus, Species, Synonyms
  \item Maximum body size (s1 - s7, Ref_s), Life cycle duration (cd1, cd2, Ref_cd), 
  Potential number of reproductive cycles per year (cy1, cy2, cy3, Ref_cy),
  Dispersal strategy (dis1 - dis4, Ref_dis), Adult life span (life1 - life4, Ref_life),
  Female wing length (mm) (fwl1 - fwl8, fwl_info, Ref_fwl),
  Wing pair type (wnb1 - wnb5, Ref_wnb), 
  Lifelong fecundity (number of eggs per female) (egg1 - egg4, Ref_egg),
  Propensity to drift (drift1, drift2, drift3, Ref_drift)
  \item scientificNameStd
}
}

