% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{elton_birds}
\alias{elton_birds}
\title{Foraging attributes of birds}
\format{
A \code{data.frame} with 9760 observations and 43 variables.
}
\source{
Cite this dataset as \itemize{ \item Hamish Wilman, Jonathan Belmaker, Jennifer Simpson, 
  Carolina de la Rosa, Marcelo M. Rivadeneira, and Walter Jetz. (2014). EltonTraits 1.0: Species-level 
  foraging attributes of the world's birds and mammals. Ecology 95:2027. \url{http://dx.doi.org/10.1890/13-1917.1}.}
}
\usage{
data(elton_birds)
}
\description{
Dataset on the foraging attributes of birds
}
\details{
This csv file contains information about the foraging attributes of
  birds from the EltonTraits dataset.
}
\section{Measures}{

 
\itemize{
  \item SpecID, PassNonPass, Order, Family, BLFamilyEnglish, BLFamSequID, Taxo, Genus, Species, English,
  \item Diet.Inv, Diet.Vend, Diet.Vect, Diet.Vfish, Diet.Vunk, Diet.Scav, Diet.Fruit, Diet.Nect, Diet.Seed, 
  Diet.PlantO, Diet.5Cat, Diet.Source, Diet.Certainty, Diet.EnteredBy, ForStrat.watbelowsurf, ForStrat.wataroundsurf, 
  ForStrat.ground, ForStrat.understory, ForStrat.midhigh, ForStrat.canopy, ForStrat.aerial, PelagicSpecialist, ForStrat.Source,
  ForStrat.SpecLevel, ForStrat.EnteredBy, Nocturnal, BodyMass.Value, BodyMass.Source, BodyMass.SpecLevel, BodyMass.Comment, 
  Record.Comment, Full.Reference,
  \item scientificNameStd
}
}

