% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{elton_mammals}
\alias{elton_mammals}
\title{Foraging attributes of mammals}
\format{
A \code{data.frame} with 5400 observations and 29 variables.
}
\source{
Cite this dataset as \itemize{ \item Hamish Wilman, Jonathan Belmaker, Jennifer Simpson, 
  Carolina de la Rosa, Marcelo M. Rivadeneira, and Walter Jetz. (2014). EltonTraits 1.0: Species-level 
  foraging attributes of the world's birds and mammals. Ecology 95:2027. \url{http://dx.doi.org/10.1890/13-1917.1}.}
}
\usage{
data(elton_mammals)
}
\description{
Dataset on the foraging attributes of mammals
}
\details{
This csv file contains information about the foraging attributes of
  mammals from the EltonTraits dataset.
}
\section{Measures}{

 
\itemize{
  \item MSW3_ID, Genus, Species, Family, 
  \item Diet.Inv, Diet.Vend, Diet.Vect, Diet.Vfish, Diet.Vunk, Diet.Scav, Diet.Fruit, Diet.Nect, Diet.Seed, Diet.PlantO, Diet.Source,
  Diet.Certainty, ForStrat.Value, ForStrat.Certainty, ForStrat.Comment, Activity.Nocturnal, Activity.Crepuscular, Activity.Diurnal, 
  Activity.Source, Activity.Certainty, BodyMass.Value, BodyMass.Source, BodyMass.SpecLevel, Full.Reference,
  \item scientificNameStd
}
}

