% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{epiphytes}
\alias{epiphytes}
\title{Vascular epiphytes traits}
\format{
A \code{data.frame} with 84109 observations and 9 variables.
}
\source{
Cite this dataset as: \itemize{\item Hietz, P., Wagner, K., Nunes Ramos, F., 
Cabral, J. S., Agudelo, C., Benavides, A. M., Cach-Pérez, M. J., Cardelús, C. L., 
Chilpa Galván, N., Erickson Nascimento da Costa, L., de Paula Oliveira, R., 
Einzmann, H. J. R., de Paiva Farias, R., Guzmán Jacob, V., Kattge, J., 
Kessler, M., Kirby, C., Kreft, H., Krömer, T., ... Zotz, G. (2021). 
Putting vascular epiphytes on the traits map. Journal of Ecology, 00, 1–19.
\url{https://doi.org/10.1111/1365-2745.13802}.}

Additionally, please cite the Dryad data package: \itemize{\item Hietz, Peter (2021)
Data from: Putting vascular epiphytes on the traits map.
Dryad Digital Repository. \url{https://doi.org/10.5061/dryad.7wm37pvtf}.}
}
\usage{
data(epiphytes)
}
\description{
Vascular epiphytes traits
}
\details{
A dataset of 76,561 trait observations for 2,882 species of vascular epiphytes.
}
\section{Measures}{


\itemize{
  \item Contributor, siteID, Individual, Genus, Species
  \item trait, trait_value, unit
  \item scientificNameStd
}
}

