% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{eubirds}
\alias{eubirds}
\title{Life-history of European birds}
\format{
A \code{data.frame} with 499 observations and 85 variables.
}
\source{
Cite this dataset as: \itemize{\item Storchova L, Horak D (2018)
Life-history characteristics of European birds. Global Ecology and Biogeography, 00, 1-7.
\url{https://doi.org/10.1111/geb.12709}.}

Additionally, please cite the Dryad data package: \itemize{\item Storchova L, Horak D (2018)
Data from: Life-history characteristics of European birds.
Dryad Digital Repository. \url{https://doi.org/10.5061/dryad.n6k3n}.}
}
\usage{
data(eubirds)
}
\description{
Life-history of European birds
}
\details{
A dataset for 499 bird species breeding in Europe and 34 key life-history traits 
represented in 85 variables. As a primary source of information we used the comprehensive 
handbook The birds of the Western Palearctic. The traits provide information about 
species-specific mean values. We did not record values for different geographical subspecies.
We chose several types of avian traits, such as morphological, reproductive 
and behavioural traits, dietary and habitat preferences.
}
\section{Measures}{


\itemize{
  \item ID, Order, Family, Genus, Species,
  \item LengthU_MEAN , WingU_MEAN, WingM_MEAN, WingF_MEAN, TailU_MEAN, TailM_MEAN, TailF_MEAN, BillU_MEAN,
  BillM_MEAN, BillF_MEAN, TarsusU_MEAN, TarsusM_MEAN, TarsusF_MEAN, WeightU_MEAN, WeightM_MEAN, WeightF_MEAN,
  Sexual.dimorphism, Clutch_MIN, Clutch_MAX, Clutch_MEAN, Broods.per.year, EggL_MEAN, EggW_MEAN, Egg_MASS, Young,
  Association.during.nesting, Nest.type, Nest.building, Mating.system, Incubation.period, Incubation.sex,
  Hatching, Eggshells, Nestling.period, Fledging.period, Parental.feeding, Age.of.independence, Feeding.independence,
  Age.of.first.breeding, Life.span, Post.fledging.mortality, Mortality.of.adults, Association.outside.the.breeding.season,
  Territoriality, Sedentary, Facultative.migrant, Short.distance.migrant, Long.distance.migrant, Deciduous.forest,
  Coniferous.forest, Woodland, Shrub, Savanna, Tundra, Grassland, Mountain.meadows, Reed, Swamps, Desert, Freshwater,
  Marine, Rocks, Human.settlements, Folivore_Y, Frugivore_Y, Granivore_Y, Arthropods_Y, Other.invertebrates_Y, Fish_Y,
  Other.vertebrates_Y, Carrion_Y, Omnivore_Y, Folivore_B, Frugivore_B, Granivore_B, Arthropods_B, Other.invertebrates_B,
  Fish_B, Other.vertebrates_B, Carrion_B, Omnivore_B, Data.source,
  \item scientificNameStd
}
}

