% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{fishmorph}
\alias{fishmorph}
\title{Morphological traits of freshwater fishes}
\format{
A \code{data.frame} with 8342 observations and 19 variables.
}
\source{
Cite this dataset as: \itemize{\item Brosse, S., Charpin, N., Su, G., 
Toussaint, A., Herrera, G., Tedesco, P.A., Villeger, S. (2021)
FISHMORPH: a global database on morphological traits of freshwater fishes. 
Global Ecology and Biogeography"", \url{https://doi.org/10.1111/geb.13395}.}
}
\usage{
data(fishmorph)
}
\description{
Morphological traits of freshwater fishes
}
\details{
The database includes ten morphological traits measured on 
10230 freshwater fish species, covering 59.71% of the world freshwater fish fauna.
}
\section{Measures}{

\itemize{
  \item Variable name, SpecCode, Superorder, Order, 
  Family, Genus species,
  \item MBI (Maximum body length), BEl (Body elongation), 
  VEp (Vertical eye position), REs (Relative eye size), 
  OGp (Oral gape position), RMl (Relative maxillary length), 
  BLs (Body lateral shape), PFv (Pectoral fin vertical position), 
  PFs (Pectoral fin size), CPt (Caudal peduncle throttling), 
  Type of illustration, Reference
  \item scientificNameStd
}
}

