% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{globalHWI}
\alias{globalHWI}
\title{Avian hand-wing index data}
\format{
A \code{data.frame} with 10338 observations and 23 variables.
}
\source{
Cite this dataset as: \itemize{\item Sheard et al. (2020) Ecological drivers of global 
gradients in avian dispersal inferred from wing morphology. Nature Communications.}
}
\usage{
data(globalHWI)
}
\description{
A global database on Avian wing morphology
}
\details{
A global dataset of avian hand-wing index (HWI), 
an estimate of wing shape widely adopted as a proxy for dispersal ability in birds
}
