% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{globTherm}
\alias{globTherm}
\title{Thermal tolerances for aquatic and terrestrial organisms}
\format{
A \code{data.frame} with 2134 observations and 47 variables.
}
\source{
Cite this dataset as: \itemize{\item Bennett, J. M. et al. (2018) GlobTherm, a global database on thermal tolerances for
aquatic and terrestrial organisms. Scientific Data. \url{https://doi.org/10.1038/sdata.2018.22}.}
}
\usage{
data(globTherm)
}
\description{
A global database on thermal tolerance for aquatic and terrestrial organisms
}
\details{
How climate affects species distributions is a longstanding question receiving renewed interest owing to the
need to predict the impacts of global warming on biodiversity. Is climate change forcing species to live near
their critical thermal limits? Are these limits likely to change through natural selection? These and other
important questions can be addressed with models relating geographical distributions of species with
climate data, but inferences made with these models are highly contingent on non-climatic factors such as
biotic interactions. Improved understanding of climate change effects on species will require extensive
analysis of thermal physiological traits, but such data are both scarce and scattered. To overcome current
limitations, we created the GlobTherm database. The database contains experimentally derived species'
thermal tolerance data currently comprising over 2000 species of terrestrial, freshwater, intertidal and
marine multicellular algae, plants, fungi, and animals. The GlobTherm database will be maintained and
curated by iDiv with the aim to keep expanding it, and enable further investigations on the effects of
climate on the distribution of life on Earth.
}
