% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{heteroptera}
\alias{heteroptera}
\title{Heteroptera morphometry traits}
\format{
A \code{data.frame} with 179 observations and 20 variables.
}
\source{
Cite this dataset as: \itemize{\item M. Gossner, Martin; K. Simons,
  Nadja; Hoeck, Leonhard; W. Weisser, Wolfgang (2016): Morphometric measures
  of Heteroptera sampled in grasslands across three regions of Germany. 
  \url{https://doi.org/10.6084/m9.figshare.c.3307611.v1}.}
  
  \url{https://figshare.com/articles/Data_Paper_Data_Paper/3561936}.
}
\usage{
data(heteroptera)
}
\description{
Morphometric measures of Heteroptera sampled in grasslands across three
regions of Germany.
}
\details{
The data sets comprises measured morphometric traits of a total of 179 Heteroptera species that were 
sampled by sweep-netting on a total of 150 managed grassland plots across three regions in Germany between 2008 and 2012. 
These plots represent the whole range of grassland management intensities from extensively used pastures to mown pastures 
to intensively managed and fertilized meadows. In this paper we provide a database of mean values of 23 morphometric measures 
across sex and morphotypes for each sampled Heteroptera species. 

Morphological traits are assumed to be related to their adaptation and function in the environment. 
Thus the relative morphometric traits can be used as proxies for ecological features of a species 
that may affect its performance or fitness. Our database can be used by future trait-based studies 
for developing and testing hypotheses of the functional significance of these traits. 

Examples include studying the functional responses of insect communities to environmental drivers or 
studying how the change in trait composition affects ecosystem processes.
}
\section{Measures}{

 
\itemize{
  \item Order, Suborder, Family, SpeciesID, Genus, Species, Author,
  \item Body_length, Body_volume, Rel_wing_length, 
  \item Hind.Femur_shape, Rel_Hind.Femur_length, Rel_Rostrum_length, Front.Femur_shape, 
  \item Body_shape, Rel_eye_size, Rel_Antenna_length, Schwaebische_Alb, Hainich, Schorfheide.Chorin,
  \item scientificNameStd
}
}

