% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{lizard_traits}
\alias{lizard_traits}
\title{Traits of lizards of the world}
\format{
A \code{data.frame} with 2692 observations and 50 variables.
}
\source{
Cite this dataset as: \itemize{\item Meiri S. (2018) Traits of lizards of the world:
Variation around a successful evolutionary design. Global Ecol Biogeogr. \url{https://doi.org/10.1111/geb.12773}.}
  
 Additionally, please cite the Dryad data package: 

\itemize{ \item Meiri, Shai (2019), Data from: Traits of lizards of the world: 
variation around a successful evolutionary design, Dryad, Dataset, \url{https://doi.org/10.5061/dryad.f6t39kj}.}
}
\usage{
data(lizard_traits)
}
\description{
Geographical, morphological, ecological, physiological and life history traits for lizards of the world.
}
\details{
The data sets comprises geographical, morphological, ecological, physiological and life history traits of 
  a total of 6657 known species of lizards.
}
