% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{mammal_diet}
\alias{mammal_diet}
\title{Mammal diet database}
\format{
A \code{data.frame} with 5374 observations and 31 variables.
}
\source{
Cite this dataset as: 

\itemize{ \item Kissling, W.D., Dalby, L., Flojgaard, C., Lenoir, J., Sandel, B., 
  Sandom, C., Trojelsgaard, K., Svenning, J. (2014). Establishing macroecological 
  trait datasets: digitalization, extrapolation, and validation of diet 
  preferences in terrestrial mammals worldwide. Ecol Evol, 4, 2913-2930. 
  \url{http://onlinelibrary.wiley.com/doi/10.1002/ece3.1136/}.}
  
Additionally, please cite the Dryad data package: 

\itemize{ \item Kissling WD, Dalby L, Flojgaard C, Lenoir J, 
Sandel B, Sandom C, Trojelsgaard K, Svenning J-C (2014) Data from: Establishing macroecological
trait datasets: digitalization, extrapolation, and validation of diet preferences in terrestrial 
mammals worldwide. Dryad Digital Repository. \url{https://doi.org/10.5061/dryad.6cd0v}.}
}
\usage{
data(mammal_diet)
}
\description{
A comprehensive global dataset of diet preferences of mammals. 
Diet information was digitized from the literature and 
extrapolated for species with missing information. The original and 
extrapolated data cover species-level diet information for >99% of all 
terrestrial mammals.
}
\details{
Ecological trait data are essential for understanding the 
  broad-scale distribution of biodiversity and its response to global change.
  For animals, diet represents a fundamental aspect of species' evolutionary 
  adaptations, ecological and functional roles, and trophic interactions. 
  However, the importance of diet for macroevolutionary and macroecological 
  dynamics remains little explored, partly because of the lack of 
  comprehensive trait datasets. We compiled and evaluated a comprehensive 
  global dataset of diet preferences of mammals. Diet 
  information was digitized from two global and cladewide data sources and 
  errors of data entry by multiple data recorders were assessed. We then 
  developed a hierarchical extrapolation procedure to fill-in diet 
  information for species with missing information. Missing data were 
  extrapolated with information from other taxonomic levels (genus, other 
  species within the same genus, or family) and this extrapolation was 
  subsequently validated both internally (with a jack-knife approach applied 
  to the compiled species-level diet data) and externally (using independent 
  species-level diet information from a comprehensive continentwide data 
  source). Finally, we grouped mammal species into trophic levels and dietary
  guilds, and their species richness as well as their proportion of total 
  richness were mapped at a global scale for those diet categories with good 
  validation results. The success rate of correctly digitizing data was 94%, 
  indicating that the consistency in data entry among multiple recorders was 
  high. Data sources provided species-level diet information for a total of 
  2033 species (38% of all 5364 terrestrial mammal species, based on the IUCN
  taxonomy). For the remaining 3331 species, diet information was mostly 
  extrapolated from genus-level diet information (48% of all terrestrial 
  mammal species), and only rarely from other species within the same genus 
  (6%) or from family level (8%). Internal and external validation showed 
  that: (1) extrapolations were most reliable for primary food items; (2) 
  several diet categories had high proportions of correctly predicted 
  diet ranks; and (3) the potential of correctly extrapolating specific diet 
  categories varied both within and among clades. Global maps of species 
  richness and proportion showed congruence among trophic levels, but also 
  substantial discrepancies between dietary guilds. MammalDIET provides a 
  comprehensive, unique and freely available dataset on diet preferences for 
  all terrestrial mammals worldwide. It enables broad-scale analyses for 
  specific trophic levels and dietary guilds, and a first assessment of trait
  conservatism in mammalian diet preferences at a global scale. The 
  digitalization, extrapolation and validation procedures could be 
  transferable to other trait data and taxa.
}
\section{Measures}{

 
\itemize{
  \item TaxonID, Order, Family, Genus, Species,
  \item Animal, Vertebrate, Mammal, Bird, Herptile, Fish, Invertebrate,
  \item Plant, Seed, Fruit, Nectar, Root, Leaf, Woody, Herbaceous, Other,
  \item TaxonomicNote, FillCode, TrophicLevel,
  \item MammalEater, Insectivore, Frugivore, Granivore, Folivore, DataSource,
  \item scientificNameStd
}
}

\author{
[Kissling, W.D.](danielkissling@web.de), Dalby, L., Flojgaard, C.,
  Lenoir, J., Sandel, B., Sandom, C., Trojelsgaard, K., Svenning, J.
}
