% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{mammal_diet2}
\alias{mammal_diet2}
\title{Mammal diet database}
\format{
A \code{data.frame} with 6625 observations and 31 variables.
}
\source{
Cite this dataset as: 

\itemize{ \item Gainsbury, Alison M.; Tallowin, Oliver J. S.; Meiri, Shai (2019), 
An updated global dataset for diet preferences in terrestrial mammals: testing the validity of extrapolation. 
Mammal Review, 48, 160-167. \url{http://doi.org/10.1111/mam.12119}.}
}
\usage{
data(mammal_diet2)
}
\description{
Combined mammal dietary dataset from Kissling et al. (2014) and the updated collected dietary information for 1261 mammalian species.
}
\details{
1. Diet is a key trait of an organism's life history that influences a broad spectrum of
ecological and evolutionary processes. Kissling et al. (2014) compiled a species-specific
dataset of diet preferences of mammals for 38% of a total of 5364 terrestrial mammalian
species assessed for the International Union for Conservation of Nature's Red List, to
facilitate future studies. The authors imputed dietary data for the remaining 62% by using
extrapolation from phylogenetic relatives. 2. We collected dietary information for 1261
mammalian species for which data were extrapolated by Kissling et al. (2014), in order to
evaluate the success with which such extrapolation can predict true diets. 3. The
extrapolation method devised by Kissling et al. (2014) performed well for broad dietary
categories (consumers of plants and animals). However, the method performed
inconsistently, and sometimes poorly, for finer dietary categories, varying in accuracy in both
dietary categories and mammalian orders. 4. The results of the extrapolation performance
serve as a cautionary tale. Given the large variation in extrapolation performance, we
recommend a more conservative approach for inferring mammalian diets, whereby dietary
extrapolation is implemented only when there is a high degree of phylogenetic conservatism
for dietary traits. Phylogenetic comparative methods can be used to detect and measure
phylogenetic signal in diet. If data for species are needed, then only the broadest feeding
categories should be used. This would ensure a greater level of accuracy and provide a more
robust dataset for further ecological and evolutionary analysis.
}
\author{
Gainsbury, Alison M., Tallowin, Oliver J. S., Meiri, Shai
}
