% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{marsupials}
\alias{marsupials}
\title{Life-history trait data of marsupials}
\format{
A \code{data.frame} with 161 observations and 77 variables.
}
\source{
Cite this dataset as: \itemize{\item Fisher et al. (2001) 
THE ECOLOGICAL BASIS OF LIFE HISTORY VARIATION IN MARSUPIALS. Ecology 82.}
}
\usage{
data(marsupials)
}
\description{
Life-history traits of marsupials
}
\details{
Variation in life history and six candidate ecological variables: type of diet, extent of intraspecific competition, 
risk of juvenile mortality, diurnal pattern of activity, arboreality, and rainfall pattern of metatherian mammals
}
