% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{migbehav_birds}
\alias{migbehav_birds}
\title{Migratory behaviour in birds}
\format{
A \code{data.frame} with 10596 observations and 31 variables.
}
\source{
Cite this dataset as: 

\itemize{ \item Eyres, A., Boehning-Gaese, K., Fritz, S.A. (2017). Quantification of climatic niches in birds: 
adding temporal dimension. Journal of Avian Biology, 000, 001-015. 
  \url{https://onlinelibrary.wiley.com/doi/10.1111/jav.01308}.}
  
Additionally, please refer to the database description at:
\url{http://dataportal-senckenberg.de/database/metacat/bikf.10058.1/bikf}.
}
\usage{
data(migbehav_birds)
}
\description{
Dataset on the timing of bird migrations
}
\details{
The dataset focuses on seasonal migration, i.e. movements causing an individual 
adult bird to be found in different locations over the course of one year, 
excluding everyday routine movements (e.g. foraging movements) 
and one-way dispersal movements by juveniles (natal dispersal) or adults (Newton 2008). 
Types or categories of migratory behaviour and different subcategories were recorded and 
classified with the Handbook of the Birds of the World (del Hoyo 1992-2013 and 
updates on the Handbook of the Birds of the World Alive website www.hbw.com, 
accessed until September 2016).
}
\section{Measures}{

 
\itemize{
  \item IOC3_1_Order, IOC3_1_Family, Genus, Species 
  \item Extinct_full, Extinct_partial, Marine_full, Marine_partial, 
  \item Migr_dir_full, Migr_dir_partial Migr_dir_local
  \item Migratory_status: directional migratory: 1609 species, 
  dispersive migratory: 482 species, nomadic: 102 species, 
  resident: 8325 species, unknown: 78 species
  \item Migratory_status_2: full directional migrant: 531 species, partial directional migrant: 1078 species, 
  full dispersive migrant: 84 species, partial dispersive migrant: 398 species, full nomad: 33 species, 
  partial nomad: 69 species, full resident: 7912 species, partial resident: 413 species, unknown: 78 species
  \item Migratory_status_3: extinct: 155 species, marine: 247 species, full directional migrant: 526 species, 
  partial directional migrant: 1019 species, full dispersive migrant: 8 species, partial dispersive migrant: 319 species, 
  full nomad: 33 species, partial nomad: 68 species, full resident: 7793 species, partial resident: 409 species, 
  unknown: 19 species
}
}

\author{
[Alison Eyres](alison.eyres@senckenberg.de), K. Boehning-Gaese, S.A. Fritz
}
