% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{pantheria}
\alias{pantheria}
\title{PanTHERIA mammal traits}
\format{
A \code{data.frame} with 5426 observations and 55 variables.
}
\source{
Cite this dataset as: 

\itemize{ \item E. Jones, Kate; Bielby, Jon; Cardillo, Marcel; A. Fritz, Susanne; 
  O'Dell, Justin; David L. Orme, C.; Safi, Kamran; Sechrest, Wes; H. Boakes, Elizabeth; Carbone, Chris; Connolly, 
  Christina; Cutts, Michael J.; Foster, Janine K.; Grenyer, Richard; Habib, Michael; Plaster, Christopher A.; 
  Price, Samantha A.; Rigby, Elizabeth A.; Rist, Janna; Teacher, Amber; Bininda-Emonds, Olaf R. P.; 
  Gittleman, John L.; M. Mace, Georgina; Purvis, Andy (2009): PanTHERIA: a species-level database of life history, 
  ecology, and geography of extant and recently extinct mammals. \url{http://esapubs.org/archive/ecol/E090/184/metadata.htm}.}
}
\usage{
data(pantheria)
}
\description{
Here we describe a global species-level data set of key
  life-history, ecological and geographical traits of all known extant and
  recently extinct mammals (PanTHERIA) developed for a number of
  macroecological and macroevolutionary research projects.
}
\details{
Data were gathered from the literature for 25 types of ecological
  and life history information for any extant or recently extinct species
  within class Mammalia (100740 data lines): 
  
  1. Activity Cycle; 2. Age at Eye Opening; 3. Age at First Birth; 4. Average
  Lifespan; 5. Body Mass; 6. Diet; 7. Dispersal Age; 8. Adult Limb Length; 9.
  Gestation Length; 10. Group Composition & Size; 11. Growth Data; 12.
  Habitat Layer; 13. Head-Body Length; 14. Interbirth Interval; 15. Litter
  size; 16. Litters Per Year; 17. Maximum Longevity; 18. Metabolic Rate; 19.
  Migratory Behaviour; 20. Mortality Data;  21. Population Density; 22.
  Ranging Behaviour; 23. Sexual
  Maturity Age; 24. Teat Number; and 25. Weaning Age.
  
  30 specific variables (see Class IV, Table 1) were extracted from the above
  data types for PanTHERIA from a total of 94729 data lines (before error
  checking). Additionally, 4 variables were derived from extracted variables
  within PanTHERIA and 19 variables were calculated from other spatial data
  sources (see Class V, Section C).
  
  see \url{http://esapubs.org/archive/ecol/E090/184/metadata.htm} for further
  information.
}
\section{Measures}{

 
\itemize{
  \item Order, Family, Genus, Species, Binomial, 
  \item ActivityCycle, AdultBodyMass_g, AdultForearmLen_mm, AdultHeadBodyLen_mm, AgeatEyeOpening_d, AgeatFirstBirth_d, 
  BasalMetRate_mLO2hr, BasalMetRateMass_g, DietBreadth, DispersalAge_d, GestationLen_d, HabitatBreadth, HomeRange_km2, 
  HomeRange_Indiv_km2, InterbirthInterval_d, LitterSize, LittersPerYear, MaxLongevity_m, NeonateBodyMass_g, NeonateHeadBodyLen_mm, 
  PopulationDensity_n.km2, PopulationGrpSize, SexualMaturityAge_d, SocialGrpSize, TeatNumber, Terrestriality, TrophicLevel, 
  WeaningAge_d, WeaningBodyMass_g, WeaningHeadBodyLen_mm, References, AdultBodyMass_g_EXT, LittersPerYear_EXT, NeonateBodyMass_g_EXT,
  WeaningBodyMass_g_EXT, GR_Area_km2, GR_MaxLat_dd, GR_MinLat_dd, GR_MidRangeLat_dd, GR_MaxLong_dd, GR_MinLong_dd, 
  GR_MidRangeLong_dd, HuPopDen_Min_n.km2, HuPopDen_Mean_n.km2, HuPopDen_5p_n.km2, HuPopDen_Change, Precip_Mean_mm, 
  Temp_Mean_01degC, AET_Mean_mm, PET_Mean_mm,  
  \item scientificNameStd
}
}

\author{
Kate E. Jones, Jon Bielby, Marcel Cardillo, Susanne A. Fritz, Justin
  O'Dell, C. David L. Orme, Kamran Safi, Wes Sechrest, Elizabeth H. Boakes,
  Chris Carbone, Christina Connolly, Michael J. Cutts, Janine K. Foster,
  Richard Grenyer, Michael Habib, Christopher A. Plaster, Samantha A. Price,
  Elizabeth A. Rigby, Janna Rist, Amber Teacher, Olaf R. P. Bininda-Emonds,
  John L. Gittleman, Georgina M. Mace, and Andy Purvis.
}
