% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{passerines}
\alias{passerines}
\title{Passerine morphology}
\format{
A \code{data.frame} with 2374 observations and 28 variables
}
\source{
Cite this dataset as: \itemize{ \item E. Ricklefs, Robert (2017): 
  Passerine morphology: external measurements of approximately one-quarter of passerine 
  bird species. Ecology, 98:1472. \href{http://doi.org/10.1002/ecy.1783}{10.1002/ecy.1783}.}
  
 The paper can be accessed, through: \url{http://onlinelibrary.wiley.com/doi/10.1002/ecy.1783/full}.
}
\usage{
data(passerines)
}
\description{
External measurements of approximately one-quarter of passerine bird species taken from Ricklefs 2017
}
\details{
The data set includes eight measurements of the external morphology of 1642 species,
roughly one-quarter of all passerine birds (Aves: Order Passeriformes), from all parts of the world, 
characterizing the relative proportions of the wing, tail, legs, and beak. 
Specimens were measured opportunistically over the past 40 years in museums in the United States and Europe.
Numbers of individuals measured per species vary from one to dozens in some cases.
Measurements for males and females of sexually size-dimorphic species are presented separately.
The measurements include total length, the lengths of the wing, tail, tarsus, and middle toe,
and the length, breadth, and depth of the beak. Particular attention was paid to obtaining a
broad representation of passerine higher taxa, with special interest in small families and subfamilies of passerines, 
as well as species produced by evolutionary radiations of birds in archipelagoes, including the Galapagos, Hawaii, and the Lesser Antilles.

Geographic distributions are summarized from Edwards's Coded List of Birds of the World.
North American and South American species are particularly well represented in the sample, 
as well as species belonging to the families  Tyrannidae, Furnariidae, Thamnophilidae, Mimidae, 
Sturnidae, Fringillidae, Parulidae, Icteridae, Cardinalidae, and Thraupidae.

The following measurement techniques, paraphrased from Ricklefs and Travis (1980) were used: 
(1) total length was measured with a plastic ruler from the tip of the bill to the tip of tail; 
(2) we measured the length of the folded wing, which was flattened along a stiff ruler, 
from the wrist to the tip of the longest primary; 
(3) length of the tail was measured from the base of the feathers in the center of the tail 
to the tip of the longest rectrix. 
We used dial calipers to measure (0.1 mm) the lengths of the 
(4) tarsus, 
(5) middle toe (to the base of the claw), and 
(6) culmen from the tip of the upper mandible to its kinetic hinge at the front of the skull, and the 
(7) depth and
(8) width of the beak at the kinetic hinge.
#' @section Measures:
 
\itemize{
  \item Length, Wing, Tail, Tarsus, Toe, `Bill L`, `Bill W`, `Bill D`,
  \item HN, N, NI, HP, E, EI, O, OI, AU, AZ, AI, Sex,
  \item Order, Family, Subfamily, Genus, Species, `TIF No.`, `IOC NO.`,
  \item scientificNameStd
}
}
\author{
[Robert E. Ricklefs](ricklefs@umsl.edu)
}
