% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{primates}
\alias{primates}
\title{Ecological traits of the world's primates}
\format{
A \code{data.frame} with 2213 observations and 49 variables.
}
\source{
Cite this dataset as: \itemize{\item Galan-Acedo et al. (2019) Ecological traits 
of the world’s primates. Scientific Data 6.}
}
\usage{
data(primates)
}
\description{
Ecological traits of the world's primates
}
\details{
A database on some important ecological traits of the world’s primates (504 species), 
including home range size, locomotion type, diel activity, trophic guild, body mass, habitat type, 
current conservation status, population trend, and geographic realm. We compiled this information 
through a careful review of 1,216 studies published between 1941 and 2018, resulting 
in a comprehensive, easily accessible and user-friendly database.
}
