% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{reptile_lifehist}
\alias{reptile_lifehist}
\title{Life-history trait database of European reptile species}
\format{
A \code{data.frame} with 1223 observations and 40 variables.
}
\source{
Cite this dataset as: \itemize{\item Grimm A, Ramirez AMP, Moulherat S, Reynaud J, Henle K (2014) 
Life-history trait database of European reptile species. Nature Conservation 9:45-67. 
\url{https://doi.org/10.3897/natureconservation.9.8908}.}
}
\usage{
data(reptile_lifehist)
}
\description{
Life-history traits of European reptile species
}
\details{
Life-history data are essential for providing answers to a wide range of questions in evolution, ecology, and
conservation biology. While life history data for many species, especially plants, are available online, life
history traits of European reptiles are available only widely scattered in different languages and primarily in printed media. 
For this reason, we generated a comprehensive trait database covering all European
reptile species. Data were compiled by searching the peer-reviewed and non-peer-reviewed literature. The
database covers the whole of Europe and neighbouring Asian and African countries. 
Traits were categorised under five main headings: Activity / Energy / Habitat; Phenology; Movement; Sexual Maturity; and
Morphometry. To ensure that the data were standardised, we defined trait data categories before we started
compiling data. All entries were checked by at least one other person. The dataset provides a unique source
for meta-analyses and modelling in ecology and conservation biology.
}
