% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_taxonomy.R
\name{standardize_taxonomy}
\alias{standardize_taxonomy}
\title{Standardize scientific names of species}
\source{
The initial code from this function was obtained from the `traitdataform` package,
which is available on GitHub: \url{https://github.com/EcologicalTraitData/traitdataform}.
}
\usage{
standardize_taxonomy(x, preferred_data_sources=c(11:15), fuzzy = TRUE, verbose = TRUE,
return=c("kingdom", "phylum", "class", "order", "family", "scientificNameStd"))
}
\arguments{
\item{x}{a \code{data.frame} containing at least the column scientificName.}

\item{preferred_data_sources}{Select the datasources that should be used for name matching, default is GBIF.}

\item{fuzzy}{if set to `FALSE` (default mode), this disables fuzzy matching
if problems with ambiguous species names arise.}

\item{verbose}{has currently no effect.}

\item{return}{a character vector containing the informatoin that should be
extracted into the output. Valid entries are the column names returned by
function `get_taxonomy()`. See 'Details'.}
}
\description{
Adds columns to a traitdata object containing accepted species
  names and relates to globally unique taxon identifiers via URI.
}
\details{
Taxonomic standardisation is an enormous challenge, as
species names constantly change and phylogenetic trees are regularly refined.

As a consequence, there is not one reference for accepted species names, as it often
depends on the field of research and taxonomic focus.

For reasons of simplicity and because of its high coverage of taxa, the
function `standardize_taxonomy()` by default uses the GBIF Backbone Taxonomy as its
reference system and resolves all provided species names to the accepted
name according to GBIF (resolving misspellings and synonyms in the
process).
}
\examples{
\dontrun{
standardize_taxonomy(x=data.frame(scientificName="Equus burchellii"))

standardize_taxonomy(x=data.frame(scientificName=c("Equus quagga")),
return=c("kingdom", "phylum", "class", "order", "family", "scientificNameStd"))
}
}
\author{
RS-eco
}
