% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{synonyms}
\alias{synonyms}
\title{Synonym names for non-standardised species}
\format{
A \code{data.frame} with 726 observations and 3 variables 
(Genus, Species and Synonym).
}
\usage{
data(synonyms)
}
\description{
Contains standardised names for species names, which could not be resolved 
by the `get_taxonomy` function.
}
\details{
This csv file contains all species names which could originally not be standardised using 
the `get_taxonomy` function, as species names were misspelled or old, and which were than manually corrected 
and checked and standardised again using the `get_taxonomy` function.
}
