% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{taxonomyStd}
\alias{taxonomyStd}
\title{Standardised species names}
\format{
A \code{data.frame} with 35052 observations and 9 variables.
}
\usage{
data(taxonomyStd)
}
\description{
Genus and species of all names contained within the dataset and their standardised scientific Name.
}
\details{
This csv file contains the genus and species of all species included in any of the datasets and
  their derived standardised scientific Name.
}
\section{Measures}{

 
\itemize{
  \item scientificName,
  \item Genus,
  \item Species,
  \item kingdom, phylum, order, class, family,
  \item scientificNameStd
}
}

