% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{tetra_density}
\alias{tetra_density}
\title{Population density estimates in terrestrial vertebrates}
\format{
A \code{data.frame} with 18246 observations and 22 variables.
}
\source{
Cite this dataset as: \itemize{\item Santini L., Isaac, N. J. B., Ficetola, G. F. (2018)
TetraDENSITY: A database of population density estimates in terrestrial vertebrates. Global Ecology and Biogeography, 27(7), 787-791.
\url{https://doi.org/10.1111/geb.12756}.}
}
\usage{
data(tetra_density)
}
\description{
Population density estimates in terrestrial vertebrates
}
\details{
A dataset for 2101 tetrapod species describing its population density at 1 degree from 1926 to 2017.
}
\section{Measures}{


\itemize{
  \item Class, Order, Family, Genus, Species, Subspecies
  \item Longitude, Latitude, Locality, Country, Year, Season/Month, Habitat, Sampling_Area, Sampling_Area_unit
  \item Density, Density_unit, Sampling_Method, Method_info, Notes, Ref_N
  \item scientificNameStd
}
}

