% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{data}
\name{trait_glossary}
\alias{trait_glossary}
\title{Trait glossary}
\format{
A \code{data.frame} with 354 observations and 8 variables.
}
\usage{
data(trait_glossary)
}
\description{
Overview of the different trait variables contained in the trait datasets
}
\details{
This csv file contains information about all trait variables 
  contained within this package.
}
\section{Measures}{

 
\itemize{
  \item datasetName
  \item columnName
  \item valueType
  \item Classes
  \item Unit 
  \item Description
  \item Comment
  \item Reference
  }
}

