% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitdata-package.R
\docType{package}
\name{traitdata}
\alias{traitdata}
\title{traitdata: Easy access to various animal trait data}
\description{
R package to access 32 different trait datasets covering numerous taxa.
}
\details{
traitdata provides an easy way to access 32 publicly available trait datasets 
(Amniota, Amphibian life-history traits, AmphiBio, Arthropods, Bird behaviour, Carabids, CLIMBER, EltonTraits, 
Eubirds, Heteroptera, Lizard traits, mammal_diet, mammal_diet2, Migratory behaviour of birds, PanTHERIA, 
Passerines, life history traits of reptiles and TetraDensity), 
which contain information on numerous taxonomic groups.

It has three main utilities:

\itemize{
\item Easy access to various trait datasets
\item Standardised taxonomic information for all data records, based on species level only
\item Standardised glossary of all datasets
\item Simple joining of multiple data sources based on their standardised taxonomic information
}

To learn more about traitdata, start with the \code{vignette("access-data")} vignette.
}
\author{
RS-eco
}
\keyword{package}
