% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomy.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Get accepted canonical names and taxonomy for a given species name}
\usage{
get_taxonomy(x, preferred_data_sources=c(11:15), fuzzy = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{a character string or vector of species names.}

\item{preferred_data_sources}{Select the datasources that should be used for name matching, default is GBIF.}

\item{fuzzy}{logical. If TRUE, function tries fuzzy matching for species requests.
May produce output, if normal matching fails. (not working!)}

\item{verbose}{logical. If FALSE, warnings and messages are suppressed.}
}
\value{
a \code{data.frame} mapping the user supplied names to the accepted taxon names and
higher taxonomic information (kingdom, phylum, class, order, family, genus).
}
\description{
Get accepted canonical names and taxonomy for a given species name
}
\details{
If 'synonym' is returned as TRUE, the user provided name has been identified as a synonym and was mapped to an accepted name.
The field confidence reports the confidence of the matching procedure performed by the function 'get_gbifid_()' of the package 'taxize'.
The taxonID is a globally valid URI that links to the taxon description of the GBIF backbone taxonomy.
}
\examples{
\dontrun{
get_taxonomy(x="Chorthippus albomarginatus")

get_taxonomy(x=c("Equus burchellii", "Equus quagga"))

get_taxonomy(x="Aegotheles tatei")
}
}
\references{
The initial code of this function was obtained from the `traitdataform` package,
which is available on GitHub: \url{https://github.com/EcologicalTraitData/traitdataform}.
}
\author{
RS-eco
}
